requirements.txt
torchmetrics
torchvision
torchinfo